Expose the coverage only if need be

+++ Earthfile
@@ -89,9 +89,11 @@ test:
 	WORKDIR /app
 	ARG test_args
  	RUN pytest ${test_args}
- 	RUN mkdir coverage && cd coverage && mv ../tests/.coverage ./ && coverage xml
-	RUN sed -r -i 's|filename=".+/site-packages/|filename="|g' coverage/coverage.xml
- 	SAVE ARTIFACT coverage /coverage
+	IF [ -e tests/.coverage]
+	   RUN mkdir coverage && cd coverage && mv ../tests/.coverage ./ && coverage xml
+ 	   RUN sed -r -i 's|filename=".+/site-packages/|filename="|g' coverage/coverage.xml
+ 	   SAVE ARTIFACT coverage /coverage
+	END
 
 coverage:
 	ARG test_args
